requirements.txt
django_environ==0.11.2
pandas==2.2.2
selenium==4.23.1
webdriver-manager==4.0.2